deleted old pulse test file and renamed new one to match the old one
