+++ test/terra/backends/test_pulse_simulator.py
@@ -12,7 +12,7 @@
 """
 PulseSimulator Integration Tests
 """
-import pdb
+
 import unittest
 from test.terra import common
 
@@ -26,7 +26,6 @@ import qiskit.pulse as pulse
 from qiskit.compiler import assemble
 from qiskit.quantum_info import state_fidelity
 
-from qiskit.test.mock.fake_openpulse_2q import FakeOpenPulse2Q
 from qiskit.pulse.commands import SamplePulse, FrameChange, PersistentValue
 from qiskit.providers.aer.openpulse.system_model import SystemModel, HamiltonianModel
 
@@ -47,389 +46,494 @@ class TestPulseSimulator(common.QiskitAerTestCase):
     """
     def setUp(self):
         """ Set configuration settings for pulse simulator """
-        # Get a pulse configuration from the mock real device
-        self.backend_mock = FakeOpenPulse2Q()
-        self.system = pulse.PulseChannelSpec.from_backend(self.backend_mock)
-        self.defaults = self.backend_mock.defaults()
-
-        # define the qubits
-        self.qubit_0 = 0
-        self.freq_qubit_0 = self.defaults.qubit_freq_est[self.qubit_0]
-
-        self.qubit_1 = 1
-        self.freq_qubit_1 = self.defaults.qubit_freq_est[self.qubit_1]
-
-        # 1q measurement map (so can measure the qubits seperately)
-        self.meas_map_1q = [[self.qubit_0], [self.qubit_1]]
-        # 2q measurement map
-        self.meas_map_2q = [[self.qubit_0, self.qubit_1]]
-
-        # define the pulse time (# of samples)
-        self.drive_samples = 100
-
-        # Define acquisition
-        acq_cmd = pulse.Acquire(duration=self.drive_samples)
-        self.acq_0 = acq_cmd(self.system.acquires[self.qubit_0],
-                             self.system.memoryslots[self.qubit_0])
-        self.acq_01 = acq_cmd(self.system.acquires, self.system.memoryslots)
-
         # Get pulse simulator backend
         self.backend_sim = qiskit.Aer.get_backend('pulse_simulator')
-
-    def single_pulse_schedule(self, phi, shape="square", gauss_sigma=0):
-        """Creates schedule for single pulse test
-        Args:
-            phi (float): drive phase (phi in Hamiltonian)
-            shape (str): shape of the pulse; defaults to square pulse
-            gauss_sigma (float): std dev for gaussian pulse if shape=="gaussian"
-        Returns:
-            schedule (pulse schedule): schedule for this test
-        """
-        # define default square drive pulse (add phase only; omega_a included in Hamiltonian)
-        const_pulse = np.ones(self.drive_samples)
-        phase = np.exp(1j * phi)
-        drive_pulse = SamplePulse(phase * const_pulse, name='drive_pulse')
-
-        # create simple Gaussian drive if set this shape
-        if shape == "gaussian":
-            times = 1.0 * np.arange(self.drive_samples)
-            gaussian = np.exp(-times**2 / 2 / gauss_sigma**2)
-            drive_pulse = SamplePulse(gaussian, name='drive_pulse')
-
-        # add commands to schedule
-        schedule = pulse.Schedule(name='drive_pulse')
-        schedule |= drive_pulse(self.system.qubits[self.qubit_0].drive)
-
-        schedule |= self.acq_0 << schedule.duration
-
-        return schedule
-
-    def frame_change_schedule(self, phi, fc_phi, dur_drive1, dur_drive2):
-        """Creates schedule for frame change test. Does a pulse w/ phase phi of duration dur_drive1,
-        then frame change of phase fc_phi, then another pulse of phase phi of duration dur_drive2.
-        The different durations for the pulses allow manipulation of rotation angles on Bloch sphere
-        Args:
-            phi (float): drive phase (phi in Hamiltonian)
-            fc_phi (float): phase for frame change
-            dur_drive1 (int): duration of first pulse
-            dur_drive2 (int): duration of second pulse
-
-        Returns:
-            schedule (pulse schedule): schedule for frame change test
-        """
-        # drive pulse (just phase; omega_a included in Hamiltonian)
-        phase = np.exp(1j * phi)
-        drive_pulse_1 = SamplePulse(phase * np.ones(dur_drive1),
-                                    name='drive_pulse_1')
-        drive_pulse_2 = SamplePulse(phase * np.ones(dur_drive2),
-                                    name='drive_pulse_2')
-
-        # frame change
-        fc_pulse = FrameChange(phase=fc_phi, name='fc')
-
-        # add commands to schedule
-        schedule = pulse.Schedule(name='fc_schedule')
-        schedule |= drive_pulse_1(self.system.qubits[self.qubit_0].drive)
-        schedule += fc_pulse(self.system.qubits[self.qubit_0].drive)
-        schedule += drive_pulse_2(self.system.qubits[self.qubit_0].drive)
-        schedule |= self.acq_0 << schedule.duration
-
-        return schedule
-
-    def persistent_value_schedule(self, omega_a_pv):
-        """Creates schedule for persistent value experiment. Creates pv pulse w/ drive amplitude
-        omega_a_pv. It does this by setting the omega_a term in the Hamiltonian = 1. Sets length of
-        the pv pulse = self.drive_samples. The product omega_a_pv*self.drive_samples, then, controls
-        the resulting state.
-        Args:
-            omega_a_pv (float): drive amplitude from the pv pulse
-
-        Returns:
-            schedule (pulse schedule): schedule for pv experiment
-        """
-        # pv pulse
-        pv_pulse = PersistentValue(value=omega_a_pv, name='pv')
-
-        # add commands to schedule
-        schedule = pulse.Schedule(name='pv_schedule')
-        schedule |= pv_pulse(self.system.qubits[self.qubit_0].drive)
-        schedule |= self.acq_0 << self.drive_samples
-
-        return schedule
-
-    def schedule_2q(self):
-        """Creates schedule for testing two qubit interaction. Specifically, do a pi pulse on qub 0
-        so it starts in the `1` state (drive channel) and then apply constant pulses to each
-        qubit (on control channel 1). This will allow us to test a swap gate.
-        Returns:
-            schedule (pulse schedule): schedule for 2q experiment
-        """
-        # set up const pulse
-        const_pulse = SamplePulse(np.ones(self.drive_samples),
-                                  name='const_pulse')
-
-        # set u channel
-        uchannel = 1  # gives omega1-omega0 (we will set equal, so don't need negation)
-
-        # add commands to schedule
-        schedule = pulse.Schedule(name='2q_schedule')
-        schedule |= const_pulse(
-            self.system.qubits[self.qubit_0].drive)  # pi pulse drive
-        schedule += const_pulse(self.system.controls[uchannel]
-                                ) << schedule.duration  # u chan pulse
-        schedule |= self.acq_01 << schedule.duration
-
-        return schedule
-
-    def create_ham_1q(self, omega_0, omega_a, qub_dim=2):
-        """Create single qubit Hamiltonian as given in class docstring
-
-        Args:
-            omega_0 (float): qubit 0 frequency
-            omega_a (float): drive amplitude
-            qub_dim (int): dimension of qubit subspace
-
-        Returns:
-            hamiltonian (dict): dictionary representation of single qubit hamiltonian
-        """
-
-        # Create the hamiltonian
-        hamiltonian = {}
-        hamiltonian['h_str'] = []
-
-        # Q0 terms
-        hamiltonian['h_str'].append('-0.5*omega0*Z0')
-        hamiltonian['h_str'].append('0.5*omegaa*X0||D0')
-
-        # Q1 terms
-        # none
-
-        # Set variables in ham
-        hamiltonian['vars'] = {'omega0': omega_0, 'omegaa': omega_a}
-
-        # set the qubit dimension to qub_dim
-        hamiltonian['qub'] = {'0': qub_dim}
-
-        return hamiltonian
-
-    def _system_model_1Q(self, omega_0, omega_a):
-
-        # make Hamiltonian
-        hamiltonian = {}
-        hamiltonian['h_str'] = ['-0.5*omega0*Z0', '0.5*omegaa*X0||D0']
-        hamiltonian['vars'] = {'omega0': omega_0, 'omegaa': omega_a}
-        hamiltonian['qub'] = {'0': 2}
-        ham_model = HamiltonianModel.from_string_spec(hamiltonian)
-
-        u_channel_lo = []
-        qubit_list = [0]
-        dt = 1.
-
-        return SystemModel(hamiltonian=ham_model,
-                           u_channel_lo=u_channel_lo,
-                           qubit_list=qubit_list,
-                           dt=dt)
-
-    def system_model_1q(self, omega_0, omega_a, qub_dim=2):
-        """Creates system_model for 1 qubit pulse simulation
-        """
-        backend_mock = FakeOpenPulse2Q()
-        backend_mock.configuration().hamiltonian = self.create_ham_1q(omega_0, omega_a, qub_dim)
-        system_model = SystemModel.from_backend(self.backend_mock, qubit_list=[0])
-        return system_model
-
-    def backend_options_1q(self):
-        backend_options = {}
-        backend_options['seed'] = 9000
-        return backend_options
-
-    def create_ham_2q(self, omega_0, omega_a, omega_i, qub_dim=2):
-        """Create two qubit Hamiltonian as given in comment of interaction test
-        Args:
-            omega_0 (float): Q0 frequency
-            omega_a (float): Q0 drive amplitude
-            omega_i (float): interaction amplitude
-            qub_dim (int): dimension of qubit subspace (same for both qubits)
-        Returns:
-            hamiltonian (dict): dictionary representation of two qubit hamiltonian
-        """
-
-        # Create the hamiltonian
-        hamiltonian = {}
-        hamiltonian['h_str'] = []
-
-        # Q0 single qubit term (used to pi pulse Q0)
-        hamiltonian['h_str'].append('-0.5*omega0*Z0')
-        hamiltonian['h_str'].append('0.5*omegaa*X0||D0')
-
-        # interaction term (uses U channels to get exponential piece)
-        hamiltonian['h_str'].append('omegai*Sp0*Sm1||U1')
-        hamiltonian['h_str'].append(
-            'omegai*Sm0*Sp1||U1')  # U1 gives diff omega_d1-omega_d0
-
-        # Set variables in ham
-        hamiltonian['vars'] = {
-            'omega0': omega_0,
-            'omegaa': omega_a,
-            'omegai': omega_i
-        }
-
-        # set the qubit dimensions to qub_dim
-        hamiltonian['qub'] = {'0': qub_dim, '1': qub_dim}
-
-        return hamiltonian
-
-    def backend_options_1q(self, omega_0, omega_a, qub_dim=2):
-        """Creates backend_options dictionary for 1 qubit pulse simulation.
-
-        Args:
-            omega_0 (float): qubit 0 frequency
-            omega_a (float): drive amplitude
-            qub_dim (int): dimension of qubit subspace
-
-        Returns:
-            dict: backend_options dictionary.
-        """
-        hamiltonian = self.create_ham_1q(omega_0, omega_a, qub_dim)
-        backend_options = self.backend_mock.configuration().to_dict()
-        backend_options['hamiltonian'] = hamiltonian
-        backend_options['qubit_list'] = [self.qubit_0]
-        backend_options['dt'] = 1.0  # makes time = self.drive_samples
-        backend_options['ode_options'] = {}  # optionally set ode settings
-        backend_options['seed'] = 9000
-        return backend_options
-
-    def backend_options_2q(self, omega_0, omega_a, omega_i, qub_dim=2):
-        """Creates backend_options dictionary for 1 qubit pulse simulation.
-
-        Args:
-            omega_0 (float): Q0 frequency
-            omega_a (float): Q0 drive amplitude
-            omega_i (float): interaction amplitude
-            qub_dim (int): dimension of qubit subspace (same for both qubits)
-
-        Returns:
-            dict: backend_options dictionary.
-        """
-        hamiltonian = self.create_ham_2q(omega_0, omega_a, omega_i, qub_dim)
-        backend_options = self.backend_mock.configuration().to_dict()
-        backend_options['hamiltonian'] = hamiltonian
-        backend_options['qubit_list'] = [self.qubit_0, self.qubit_1]
-        backend_options['dt'] = 1.0  # makes time = self.drive_samples
-        backend_options['ode_options'] = {}  # optionally set ode settings
-        backend_options['seed'] = 12387
-        return backend_options
-
-    def qobj_params_1q(self, omega_d0):
-        """Set params needed to create qobj for 1q tests
-        Args:
-            omega_d0 (float): qubit 0 drive (lo) frequency
-        Returns:
-            dict: contains memory_slots, qubit_lo_freq, meas_map for 1q qobj
-        """
-        memory_slots = 1
-        qubit_lo_freq = [omega_d0 / (2 * np.pi)]
-        meas_map = self.meas_map_1q
-
-        return (memory_slots, qubit_lo_freq, meas_map)
-
-    def qobj_params_2q(self, omega_d0, omega_d1):
-        """Set params needed to create qobj for 2q tests
-        Args:
-            omega_d0 (float): qubit 0 drive (lo) frequency
-            omega_d1 (float): qubit 1 drive (lo) frequency
-         Returns:
-            dict: contains memory_slots, qubit_lo_freq, meas_map for 2q qobj
-        """
-        memory_slots = 2
-        qubit_lo_freq = [omega_d0 / (2 * np.pi), omega_d1 / (2 * np.pi)]
-        meas_map = self.meas_map_2q
-
-        return (memory_slots, qubit_lo_freq, meas_map)
-
-    def create_qobj(self, shots, meas_level, schedule, qobj_params):
-        """Creates qobj for the specified pulse experiment. Uses Hamiltonian from class docstring
-        (except for 2q tests, which use Hamiltonian specified in comment of that test section).
-        Args:
-            shots (int): number of times to perform experiment
-            meas_level (int): level of data to return
-            schedule (Schedule): pulse schedule for the qobj
-            qobj_params (tuple): tuple of memory_slots, qubit_lo_freq, meas_map
-        Returns:
-            Qobj: qobj representing this pulse experiment
-        """
-        # set qobj params
-        memory_slots = qobj_params[0]
-        qubit_lo_freq = qobj_params[1]
-        meas_map = qobj_params[2]
-
-        # construct the qobj
-        qobj = assemble([schedule],
-                        self.backend_mock,
-                        meas_level=meas_level,
-                        meas_return='single',
-                        meas_map=meas_map,
-                        qubit_lo_freq=qubit_lo_freq,
-                        memory_slots=memory_slots,
-                        shots=shots)
-
-        return qobj
 
     # ---------------------------------------------------------------------
     # Test single qubit gates (using meas level 2 and square drive)
     # ---------------------------------------------------------------------
 
+    def test_x_gate(self):
+        """
+        Test x gate. Set omega_d0=omega_0 (drive on resonance), phi=0, omega_a = pi/time
+        """
+
+        # setup system model
+        total_samples = 100
+        omega_0 = 2 * np.pi
+        omega_d0 = omega_0
+        omega_a = np.pi / total_samples
+        system_model = self._system_model_1Q(omega_0, omega_a)
+
+        # set up schedule and qobj
+        schedule = self._simple_1Q_schedule(system_model, 0, total_samples)
+        qobj = assemble([schedule],
+                        meas_level=2,
+                        meas_return='single',
+                        meas_map=[[0]],
+                        qubit_lo_freq=[omega_d0/(2*np.pi)],
+                        meas_lo_freq=[0.],
+                        memory_slots=2,
+                        shots=256)
+
+        # set backend backend_options
+        backend_options = {'seed' : 9000}
+
+        # run simulation
+        result = self.backend_sim.run(qobj, system_model=system_model,
+                                      backend_options=backend_options).result()
+
+        # test results
+        counts = result.get_counts()
+        exp_counts = {'1': 256}
+        self.assertDictAlmostEqual(counts, exp_counts)
+
+    def test_dt_scaling_x_gate(self):
+        """
+        Test that dt is being used correctly by the solver.
+        """
+
+        total_samples = 100
+        # do the same thing as test_x_gate, but scale dt and all frequency parameters
+        # define test case for a single scaling
+        def scale_test(scale):
+            # set omega_0, omega_d0 equal (use qubit frequency) -> drive on resonance
+            # Require omega_a*time = pi to implement pi pulse (x gate)
+            omega_0 = 2 * np.pi / scale
+            omega_d0 = omega_0
+            omega_a = np.pi / total_samples / scale
+
+            # set up system model
+            system_model = self._system_model_1Q(omega_0, omega_a)
+            system_model.dt = system_model.dt * scale
+
+            # set up schedule and qobj
+            schedule = self._simple_1Q_schedule(system_model, 0, total_samples)
+            qobj = assemble([schedule],
+                            meas_level=2,
+                            meas_return='single',
+                            meas_map=[[0]],
+                            qubit_lo_freq=[omega_d0/(2*np.pi)],
+                            meas_lo_freq=[0.],
+                            memory_slots=2,
+                            shots=256)
+
+            # set backend backend_options
+            backend_options = {'seed' : 9000}
+
+            # run simulation
+            result = self.backend_sim.run(qobj, system_model=system_model,
+                                          backend_options=backend_options).result()
+            counts = result.get_counts()
+            exp_counts = {'1': 256}
+
+            self.assertDictAlmostEqual(counts, exp_counts)
+
+        # set scales and run tests
+        scales = [2., 0.1234, 10.**5, 10**-5]
+        for scale in scales:
+            scale_test(scale)
+
+    def test_hadamard_gate(self):
+        """Test Hadamard. Is a rotation of pi/2 about the y-axis. Set omega_d0=omega_0
+        (drive on resonance), phi=-pi/2, omega_a = pi/2/time
+        """
+
+        # set variables
+        shots = 100000  # large number of shots so get good proportions
+        total_samples = 100
+
+        # set omega_0, omega_d0 equal (use qubit frequency) -> drive on resonance
+        omega_0 = 2 * np.pi
+        omega_d0 = omega_0
+
+        # Require omega_a*time = pi/2 to implement pi/2 rotation pulse
+        # num of samples gives time
+        omega_a = np.pi / 2 / total_samples
+
+        system_model = self._system_model_1Q(omega_0, omega_a)
+
+        phi = -np.pi / 2
+        schedule = self._simple_1Q_schedule(system_model, phi, total_samples)
+
+        qobj = assemble([schedule],
+                        meas_level=2,
+                        meas_return='single',
+                        meas_map=[[0]],
+                        qubit_lo_freq=[omega_d0/(2*np.pi)],
+                        meas_lo_freq=[0.],
+                        memory_slots=2,
+                        shots=shots)
+
+        # set backend backend_options
+        backend_options = {'seed' : 9000}
+
+        # run simulation
+        result = self.backend_sim.run(qobj, system_model=system_model,
+                                      backend_options=backend_options).result()
+        counts = result.get_counts()
+
+        # compare prop
+        prop = {}
+        for key in counts.keys():
+            prop[key] = counts[key] / shots
+
+        exp_prop = {'0': 0.5, '1': 0.5}
+
+        self.assertDictAlmostEqual(prop, exp_prop, delta=0.01)
+
+    def test_arbitrary_gate(self):
+        """Test a few examples w/ arbitary drive, phase and amplitude. """
+        shots = 10000  # large number of shots so get good proportions
+        total_samples = 100
+        num_tests = 3
+        # set variables for each test
+        omega_0 = 2 * np.pi
+        omega_d0_vals = [omega_0 + 1, omega_0 + 0.02, omega_0 + 0.005]
+        omega_a_vals = [
+            2 * np.pi / 3 / total_samples,
+            7 * np.pi / 5 / total_samples, 0.1
+        ]
+        phi_vals = [5 * np.pi / 7, 19 * np.pi / 14, np.pi / 4]
+
+        for i in range(num_tests):
+            with self.subTest(i=i):
+
+                system_model = self._system_model_1Q(omega_0, omega_a_vals[i])
+                schedule = self._simple_1Q_schedule(system_model, phi_vals[i], total_samples)
+
+                qobj = assemble([schedule],
+                                meas_level=2,
+                                meas_return='single',
+                                meas_map=[[0]],
+                                qubit_lo_freq=[omega_d0_vals[i]/(2*np.pi)],
+                                meas_lo_freq=[0.],
+                                memory_slots=2,
+                                shots=shots)
+
+                # Run qobj and compare prop to expected result
+                backend_options = {'seed' : 9000}
+                result = self.backend_sim.run(qobj, system_model, backend_options).result()
+                counts = result.get_counts()
+
+                prop = {}
+                for key in counts.keys():
+                    prop[key] = counts[key] / shots
+
+                exp_prop = self._analytic_prop_1q_gates(
+                    total_samples=total_samples,
+                    omega_0=omega_0,
+                    omega_a=omega_a_vals[i],
+                    omega_d0=omega_d0_vals[i],
+                    phi=phi_vals[i])
+
+                self.assertDictAlmostEqual(prop, exp_prop, delta=0.01)
+
+    def test_meas_level_1(self):
+        """Test measurement level 1. """
+
+        shots = 10000  # run large number of shots for good proportions
+        total_samples = 100
+        # perform hadamard setup (so get some 0's and some 1's), but use meas_level = 1
+
+        # set omega_0, omega_d0 equal (use qubit frequency) -> drive on resonance
+        omega_0 = 2 * np.pi
+        omega_d0 = omega_0
+
+        # Require omega_a*time = pi/2 to implement pi/2 rotation pulse
+        # num of samples gives time
+        omega_a = np.pi / 2 / total_samples
+
+        phi = -np.pi / 2
+
+        system_model = self._system_model_1Q(omega_0, omega_a)
+
+        phi = -np.pi / 2
+        schedule = self._simple_1Q_schedule(system_model, phi, total_samples)
+
+        qobj = assemble([schedule],
+                        meas_level=1,
+                        meas_return='single',
+                        meas_map=[[0]],
+                        qubit_lo_freq=[omega_d0/(2*np.pi)],
+                        meas_lo_freq=[0.],
+                        memory_slots=2,
+                        shots=shots)
+
+        # set backend backend_options
+        backend_options = {'seed' : 9000}
+
+        result = self.backend_sim.run(qobj, system_model, backend_options).result()
+
+        # Verify that (about) half the IQ vals have abs val 1 and half have abs val 0
+        # (use prop for easier comparison)
+        mem = np.abs(result.get_memory()[:, 0])
+
+        iq_prop = {'0': 0, '1': 0}
+        for i in mem:
+            if i == 0:
+                iq_prop['0'] += 1 / shots
+            else:
+                iq_prop['1'] += 1 / shots
+
+        exp_prop = {'0': 0.5, '1': 0.5}
+
+        self.assertDictAlmostEqual(iq_prop, exp_prop, delta=0.01)
+
+    def test_gaussian_drive(self):
+        """Test gaussian drive pulse using meas_level_2. Set omega_d0=omega_0 (drive on resonance),
+        phi=0, omega_a = pi/time
+        """
+
+        # set variables
+
+        # set omega_0, omega_d0 equal (use qubit frequency) -> drive on resonance
+        total_samples = 100
+        omega_0 = 2 * np.pi
+        omega_d0 = omega_0
+
+        # Require omega_a*time = pi to implement pi pulse (x gate)
+        # num of samples gives time
+        omega_a = np.pi / total_samples
+
+        phi = 0
+
+        # Test gaussian drive results for a few different sigma
+        gauss_sigmas = {
+            total_samples / 6, total_samples / 3, total_samples
+        }
+
+        system_model = self._system_model_1Q(omega_0, omega_a)
+
+        for gauss_sigma in gauss_sigmas:
+            with self.subTest(gauss_sigma=gauss_sigma):
+                schedule = self._simple_1Q_schedule(system_model,
+                                                    phi,
+                                                    total_samples,
+                                                    shape="gaussian",
+                                                    gauss_sigma=gauss_sigma)
+
+                qobj = assemble([schedule],
+                                meas_level=2,
+                                meas_return='single',
+                                meas_map=[[0]],
+                                qubit_lo_freq=[omega_d0/(2*np.pi)],
+                                meas_lo_freq=[0.],
+                                memory_slots=2,
+                                shots=1000)
+                backend_options = {'seed' : 9000}
+
+                result = self.backend_sim.run(qobj, system_model, backend_options).result()
+                statevector = result.get_statevector()
+                exp_statevector = self._analytic_gaussian_statevector(
+                    total_samples, gauss_sigma=gauss_sigma, omega_a=omega_a)
+
+                # Check fidelity of statevectors
+                self.assertGreaterEqual(
+                    state_fidelity(statevector, exp_statevector), 0.99)
+
+    def test_frame_change(self):
+        """Test frame change command. """
+        shots = 10000
+        total_samples = 100
+        # set omega_0, omega_d0 equal (use qubit frequency) -> drive on resonance
+        omega_0 = 2 * np.pi
+        omega_d0 = omega_0
+
+        # set phi = 0
+        phi = 0
+
+        dur_drive1 = total_samples  # first pulse duration
+        fc_phi = np.pi
+
+        # Test frame change where no shift in state results
+        # specfically: do pi/2 pulse, then pi frame change, then another pi/2 pulse.
+        # Verify left in |0> state
+        dur_drive2 = dur_drive1  # same duration for both pulses
+        omega_a = np.pi / 2 / dur_drive1  # pi/2 pulse amplitude
+
+        system_model = self._system_model_1Q(omega_0, omega_a)
+        schedule = self._1Q_frame_change_schedule(system_model,
+                                                  phi,
+                                                  fc_phi,
+                                                  total_samples,
+                                                  dur_drive1,
+                                                  dur_drive2)
+        qobj = assemble([schedule],
+                        meas_level=2,
+                        meas_return='single',
+                        meas_map=[[0]],
+                        qubit_lo_freq=[omega_d0/(2*np.pi)],
+                        meas_lo_freq=[0.],
+                        memory_slots=2,
+                        shots=shots)
+
+        backend_options = {'seed' : 9000}
+        result = self.backend_sim.run(qobj, system_model, backend_options).result()
+        counts = result.get_counts()
+        exp_counts = {'0': shots}
+
+        self.assertDictAlmostEqual(counts, exp_counts)
+
+        # Test frame change where a shift does result
+        # specifically: do pi/4 pulse, then pi phase change, then do pi/8 pulse.
+        # check that a net rotation of pi/4-pi/8 has occured on the Bloch sphere
+        dur_drive2 = int(dur_drive1 / 2)  # half time for second pulse (halves angle)
+        omega_a = np.pi / 4 / dur_drive1  # pi/4 pulse amplitude
+
+        system_model = self._system_model_1Q(omega_0, omega_a)
+        schedule = self._1Q_frame_change_schedule(system_model,
+                                                  phi,
+                                                  fc_phi,
+                                                  total_samples,
+                                                  dur_drive1,
+                                                  dur_drive2)
+        qobj = assemble([schedule],
+                        meas_level=2,
+                        meas_return='single',
+                        meas_map=[[0]],
+                        qubit_lo_freq=[omega_d0/(2*np.pi)],
+                        meas_lo_freq=[0.],
+                        memory_slots=2,
+                        shots=shots)
+
+        backend_options = {'seed' : 9000}
+        result = self.backend_sim.run(qobj, system_model, backend_options).result()
+        counts = result.get_counts()
+
+        # verify props
+        prop_shift = {}
+        for key in counts.keys():
+            prop_shift[key] = counts[key] / shots
+
+        # net angle is given by pi/4-pi/8
+        prop0 = np.cos( (np.pi / 4 - np.pi / 8) / 2)**2
+        exp_prop = {'0' : prop0, '1': 1 - prop0}
+        self.assertDictAlmostEqual(prop_shift, exp_prop, delta=0.01)
+
+    def test_three_level(self):
+        r"""Test 3 level system. Compare statevectors as counts only use bitstrings. Analytic form
+        given in _analytic_statevector_3level function docstring.
+        """
+
+        def analytic_state_vector(omega_a, total_samples):
+            r"""Returns analytically computed statevector for 3 level system with our Hamiltonian. Is
+            given by `(\frac{1}{3} (2+\cos(\frac{\sqrt{3}}{2} \omega_a t)),
+            -\frac{i}{\sqrt{3}} \sin(\frac{\sqrt{3}}{2} \omega_a t),
+            -\frac{2\sqrt{2}}{3} \sin(\frac{\sqrt{3}}{4} \omega_a t)^2)`.
+            Args:
+                omega_a (float): Q0 drive amplitude
+            Returns:
+                exp_statevector (list): analytically computed statevector with Hamiltonian from above
+                    (Returned in the rotating frame)
+            """
+            time = total_samples
+            arg1 = np.sqrt(3) * omega_a * time / 2  # cos arg for first component
+            arg2 = arg1  # sin arg for first component
+            arg3 = arg1 / 2  # sin arg for 3rd component
+            exp_statevector = np.array([(2 + np.cos(arg1)) / 3,
+                                        -1j * np.sin(arg2) / np.sqrt(3),
+                                        -2 * np.sqrt(2) * np.sin(arg3)**2 / 3],
+                                        dtype=complex)
+            return exp_statevector
 
 
-    # ---------------------------------------------------------------------
-    # Test FrameChange and PersistentValue commands
-    # ---------------------------------------------------------------------
+        shots = 1000
+        total_samples = 100
+        # Set omega_0,omega_d0 (use qubit frequency) -> drive on resonance
+        omega_0 = 2 * np.pi
+        omega_d0 = omega_0
 
+        # Set phi = 0 for simplicity
+        phi = 0
 
-    # ----------------------------------------------------------------------------------------------
-    # Test qubit interaction (use 2 qubits for simplicity)
-    # For these tests, we use a different 2-qubit Hamiltonian that tests both
-    # interaction and control (U) channels. In the lab frame, it is given
-    # by `H = -\frac{1}{2} \omega_0 \sigma_z^0 + \frac{1}{2} \omega_a e^{i \omega_{d0} t} \sigma_x^0
-    # `+ \omega_i (e^{i (\omega_{d0}-\omega_{d1}) t} \sigma_{p0} \otimes \sigma_{m1} + `
-    # `+ e^{-i (\omega_{d0}-\omega_{d1}) t} \sigma_{m0} \otimes \sigma_{p1})`. First 2 terms allow
-    # us to excite the 0 qubit. Latter 2 terms define the interaction.
-    # ----------------------------------------------------------------------------------------------
+        # Test pi pulse
+        omega_a = np.pi / total_samples
+
+        system_model = self._system_model_1Q(omega_0, omega_a, dim_qub=3)
+        schedule = self._simple_1Q_schedule(system_model, phi, total_samples)
+
+        qobj = assemble([schedule],
+                        meas_level=2,
+                        meas_return='single',
+                        meas_map=[[0]],
+                        qubit_lo_freq=[omega_d0/(2*np.pi)],
+                        meas_lo_freq=[0.],
+                        memory_slots=2,
+                        shots=shots)
+        backend_options = {'seed' : 9000}
+
+        result = self.backend_sim.run(qobj, system_model, backend_options).result()
+        statevector = result.get_statevector()
+
+        exp_statevector = analytic_state_vector(omega_a, total_samples)
+
+        # Check fidelity of statevectors
+        self.assertGreaterEqual(
+            state_fidelity(statevector, exp_statevector), 0.99)
+
+        # Test 2*pi pulse
+        omega_a = 2 * np.pi / total_samples
+
+        system_model = self._system_model_1Q(omega_0, omega_a, dim_qub=3)
+        schedule = self._simple_1Q_schedule(system_model, phi, total_samples)
+
+        qobj = assemble([schedule],
+                        meas_level=2,
+                        meas_return='single',
+                        meas_map=[[0]],
+                        qubit_lo_freq=[omega_d0/(2*np.pi)],
+                        meas_lo_freq=[0.],
+                        memory_slots=2,
+                        shots=shots)
+        backend_options = {'seed' : 9000}
+
+        result = self.backend_sim.run(qobj, system_model, backend_options).result()
+        statevector = result.get_statevector()
+
+        exp_statevector = analytic_state_vector(omega_a, total_samples)
+
+        # Check fidelity of vectors
+        self.assertGreaterEqual(
+            state_fidelity(statevector, exp_statevector), 0.99)
 
     def test_interaction(self):
         r"""Test 2 qubit interaction via swap gates."""
 
         shots = 100000
-
+        total_samples = 100
         # Do a standard SWAP gate
 
         # Interaction amp (any non-zero creates the swap gate)
-        omega_i_swap = np.pi / 2 / self.drive_samples
+        omega_i_swap = np.pi / 2 / total_samples
         # set omega_d0=omega_0 (resonance)
-        omega_0 = 2 * np.pi * self.freq_qubit_0
+        omega_0 = 2 * np.pi
         omega_d0 = omega_0
 
         # For swapping, set omega_d1 = 0 (drive on Q0 resonance)
-        omega_d1_swap = 0
+        # Note: confused by this as there is no d1 term
+        omega_d1 = 0
 
         # do pi pulse on Q0 and verify state swaps from '01' to '10' (reverse bit order)
 
         # Q0 drive amp -> pi pulse
-        omega_a_pi_swap = np.pi / self.drive_samples
+        omega_a_pi_swap = np.pi / total_samples
 
-        schedule_pi_swap = self.schedule_2q()
-        qobj_params_pi_swap = self.qobj_params_2q(omega_d0=omega_d0,
-                                                  omega_d1=omega_d1_swap)
 
-        qobj_pi_swap = self.create_qobj(shots=shots,
-                                        meas_level=2,
-                                        schedule=schedule_pi_swap,
-                                        qobj_params=qobj_params_pi_swap)
+        system_model = self._system_model_2Q(omega_0, omega_a_pi_swap, omega_i_swap)
 
-        backend_options_pi_swap = self.backend_options_2q(
-            omega_0, omega_a_pi_swap, omega_i_swap)
-        result_pi_swap = self.backend_sim.run(
-            qobj_pi_swap, backend_options=backend_options_pi_swap).result()
+        schedule = self._schedule_2Q_interaction(system_model, total_samples)
+        qobj = assemble([schedule],
+                        meas_level=2,
+                        meas_return='single',
+                        meas_map=[[0, 1]],
+                        qubit_lo_freq=[omega_d0 / (2 * np.pi), omega_d1 / (2 * np.pi)],
+                        meas_lo_freq=[0., 0.],
+                        memory_slots=2,
+                        shots=shots)
+        backend_options ={'seed': 12387}
+
+        result_pi_swap = self.backend_sim.run(qobj, system_model, backend_options).result()
         counts_pi_swap = result_pi_swap.get_counts()
 
         exp_counts_pi_swap = {
@@ -440,21 +544,12 @@ class TestPulseSimulator(common.QiskitAerTestCase):
         # do pi/2 pulse on Q0 and verify half the counts are '00' and half are swapped state '10'
 
         # Q0 drive amp -> pi/2 pulse
-        omega_a_pi2_swap = np.pi / 2 / self.drive_samples
+        omega_a_pi2_swap = np.pi / 2 / total_samples
 
-        schedule_pi2_swap = self.schedule_2q()
-        qobj_params_pi2_swap = self.qobj_params_2q(omega_d0=omega_d0,
-                                                   omega_d1=omega_d1_swap)
+        # update system model with this new variable
+        system_model.hamiltonian.set_vars({'omegaa' : omega_a_pi2_swap})
 
-        qobj_pi2_swap = self.create_qobj(shots=shots,
-                                         meas_level=2,
-                                         schedule=schedule_pi2_swap,
-                                         qobj_params=qobj_params_pi2_swap)
-
-        backend_options_pi2_swap = self.backend_options_2q(
-            omega_0, omega_a_pi2_swap, omega_i_swap)
-        result_pi2_swap = self.backend_sim.run(
-            qobj_pi2_swap, backend_options=backend_options_pi2_swap).result()
+        result_pi2_swap = self.backend_sim.run(qobj, system_model, backend_options).result()
         counts_pi2_swap = result_pi2_swap.get_counts()
 
         # compare proportions for improved accuracy
@@ -475,27 +570,207 @@ class TestPulseSimulator(common.QiskitAerTestCase):
         omega_d1_no_swap = omega_d0
 
         # Q0 drive amp -> pi pulse
-        omega_a_no_swap = np.pi / self.drive_samples
+        omega_a_no_swap = np.pi / total_samples
 
-        schedule_no_swap = self.schedule_2q()
-        qobj_params_no_swap = self.qobj_params_2q(omega_d0=omega_d0,
-                                                  omega_d1=omega_d1_no_swap)
+        # update system model with this new variable
+        system_model.hamiltonian.set_vars({'omega0' : omega_0,
+                                           'omegaa' : omega_a_no_swap,
+                                           'omegai' : omega_i_no_swap})
 
-        qobj_no_swap = self.create_qobj(shots=shots,
-                                        meas_level=2,
-                                        schedule=schedule_no_swap,
-                                        qobj_params=qobj_params_no_swap)
-        backend_options_swap = self.backend_options_2q(omega_0,
-                                                       omega_a_no_swap,
-                                                       omega_i_no_swap)
-        result_no_swap = self.backend_sim.run(
-            qobj_no_swap, backend_options=backend_options_swap).result()
+        result_no_swap = self.backend_sim.run(qobj, system_model, backend_options).result()
         counts_no_swap = result_no_swap.get_counts()
 
         exp_counts_no_swap = {
             '01': shots
         }  # non-swapped state (reverse bit order)
         self.assertDictAlmostEqual(counts_no_swap, exp_counts_no_swap)
+
+
+
+    def _system_model_1Q(self, omega_0, omega_a, dim_qub=2):
+
+        # make Hamiltonian
+        hamiltonian = {}
+        hamiltonian['h_str'] = ['-0.5*omega0*Z0', '0.5*omegaa*X0||D0']
+        hamiltonian['vars'] = {'omega0': omega_0, 'omegaa': omega_a}
+        hamiltonian['qub'] = {'0': dim_qub}
+        ham_model = HamiltonianModel.from_string_spec(hamiltonian)
+
+        u_channel_lo = []
+        qubit_list = [0]
+        dt = 1.
+
+        return SystemModel(hamiltonian=ham_model,
+                           u_channel_lo=u_channel_lo,
+                           qubit_list=qubit_list,
+                           dt=dt)
+
+    def _simple_1Q_schedule(self, system_model, phi, total_samples, shape="square", gauss_sigma=0):
+        """Creates schedule for single pulse test
+        Args:
+            phi (float): drive phase (phi in Hamiltonian)
+            shape (str): shape of the pulse; defaults to square pulse
+            gauss_sigma (float): std dev for gaussian pulse if shape=="gaussian"
+        Returns:
+            schedule (pulse schedule): schedule for this test
+        """
+
+        # set up pulse command
+        phase = np.exp(1j * phi)
+        drive_pulse = None
+        if shape == "square":
+            const_pulse = np.ones(total_samples)
+            drive_pulse = SamplePulse(phase * const_pulse, name='drive_pulse')
+        if shape == "gaussian":
+            times = 1.0 * np.arange(total_samples)
+            gaussian = np.exp(-times**2 / 2 / gauss_sigma**2)
+            drive_pulse = SamplePulse(phase * gaussian, name='drive_pulse')
+
+        # set up acquire command
+        acq_cmd = pulse.Acquire(duration=total_samples)
+
+        # add commands into a schedule for first qubit
+        schedule = pulse.Schedule(name='drive_pulse')
+        schedule |= drive_pulse(system_model.drive(0))
+        schedule |= acq_cmd(system_model.acquire(0),
+                            system_model.memoryslot(0)) << schedule.duration
+
+        return schedule
+
+    def _1Q_frame_change_schedule(self,
+                                  system_model,
+                                  phi,
+                                  fc_phi,
+                                  total_samples,
+                                  dur_drive1,
+                                  dur_drive2):
+        """Creates schedule for frame change test. Does a pulse w/ phase phi of duration dur_drive1,
+        then frame change of phase fc_phi, then another pulse of phase phi of duration dur_drive2.
+        The different durations for the pulses allow manipulation of rotation angles on Bloch sphere
+        Args:
+            phi (float): drive phase (phi in Hamiltonian)
+            fc_phi (float): phase for frame change
+            dur_drive1 (int): duration of first pulse
+            dur_drive2 (int): duration of second pulse
+
+        Returns:
+            schedule (pulse schedule): schedule for frame change test
+        """
+        phase = np.exp(1j * phi)
+        drive_pulse_1 = SamplePulse(phase * np.ones(dur_drive1),
+                                    name='drive_pulse_1')
+        drive_pulse_2 = SamplePulse(phase * np.ones(dur_drive2),
+                                    name='drive_pulse_2')
+
+        # frame change
+        fc_pulse = FrameChange(phase=fc_phi, name='fc')
+
+        # set up acquire command
+        acq_cmd = pulse.Acquire(duration=total_samples)
+
+        # add commands to schedule
+        schedule = pulse.Schedule(name='fc_schedule')
+        schedule |= drive_pulse_1(system_model.drive(0))
+        schedule += fc_pulse(system_model.drive(0))
+        schedule += drive_pulse_2(system_model.drive(0))
+        schedule |= acq_cmd(system_model.acquire(0),
+                            system_model.memoryslot(0)) << schedule.duration
+
+        return schedule
+
+    def _analytic_prop_1q_gates(self, total_samples, omega_0, omega_a, omega_d0, phi):
+        """Compute proportion for 0 and 1 states analytically for single qubit gates.
+        Args:
+            omega_0 (float): Q0 freq
+            omega_a (float): Q0 drive amplitude
+            omega_d0 (flaot): Q0 drive frequency
+            phi (float): drive phase
+        Returns:
+            exp_prop (dict): expected value of 0 and 1 proportions from analytic computation
+            """
+        time = total_samples
+        # write Hrot analytically
+        h_rot = np.array([[
+            (omega_d0 - omega_0) / 2,
+            np.exp(1j * phi) * omega_a / 2
+        ], [np.exp(-1j * phi) * omega_a / 2, -(omega_d0 - omega_0) / 2]])
+        # exponentiate
+        u_rot = expm(-1j * h_rot * time)
+        state0 = np.array([1, 0])
+
+        # compute analytic prob (proportion) of 0 state
+        mat_elem0 = np.vdot(state0, np.dot(u_rot, state0))
+        prop0 = np.abs(mat_elem0)**2
+
+        # return expected proportion
+        exp_prop = {'0': prop0, '1': 1 - prop0}
+        return exp_prop
+
+    def _analytic_gaussian_statevector(self, total_samples, gauss_sigma, omega_a):
+        r"""Computes analytic statevector for gaussian drive. Solving the Schrodinger equation in
+        the rotating frame leads to the analytic solution `(\cos(x), -i\sin(x)) with
+        `x = \frac{1}{2}\sqrt{\frac{\pi}{2}}\sigma\omega_a erf(\frac{t}{\sqrt{2}\sigma}).
+        Args:
+            gauss_sigma (float): std dev for the gaussian drive
+            omega_a (float): Q0 drive amplitude
+        Returns:
+            exp_statevector (list): analytic form of the statevector computed for gaussian drive
+                (Returned in the rotating frame)
+        """
+        time = total_samples
+        arg = 1 / 2 * np.sqrt(np.pi / 2) * gauss_sigma * omega_a * erf(
+            time / np.sqrt(2) / gauss_sigma)
+        exp_statevector = [np.cos(arg), -1j * np.sin(arg)]
+        return exp_statevector
+
+    def _system_model_2Q(self, omega_0, omega_a, omega_i, dim_qub=2):
+        # make Hamiltonian
+        hamiltonian = {}
+        # qubit 0 terms
+        hamiltonian['h_str'] = ['-0.5*omega0*Z0', '0.5*omegaa*X0||D0']
+        # interaction term
+        hamiltonian['h_str'].append('omegai*(Sp0*Sm1+Sm0*Sp1)||U1')
+        hamiltonian['vars'] = {
+            'omega0': omega_0,
+            'omegaa': omega_a,
+            'omegai': omega_i
+        }
+        hamiltonian['qub'] = {'0' : dim_qub, '1' : dim_qub}
+        ham_model = HamiltonianModel.from_string_spec(hamiltonian)
+
+
+        u_channel_lo = [ [{'q': 0, 'scale': [1.0, 0.0]}],
+                         [{'q': 0, 'scale': [-1.0, 0.0]}, {'q': 1, 'scale': [1.0, 0.0]}] ]
+        qubit_list = [0, 1]
+        dt = 1.
+
+        return SystemModel(hamiltonian=ham_model,
+                           u_channel_lo=u_channel_lo,
+                           qubit_list=qubit_list,
+                           dt=dt)
+
+    def _schedule_2Q_interaction(self, system_model, total_samples):
+        """Creates schedule for testing two qubit interaction. Specifically, do a pi pulse on qub 0
+        so it starts in the `1` state (drive channel) and then apply constant pulses to each
+        qubit (on control channel 1). This will allow us to test a swap gate.
+        Returns:
+            schedule (pulse schedule): schedule for 2q experiment
+        """
+        # set up const pulse
+        const_pulse = SamplePulse(np.ones(total_samples), name='const_pulse')
+
+        # set u channel
+        uchannel = 1  # gives omega1-omega0 (we will set equal, so don't need negation)
+
+        # add commands to schedule
+        schedule = pulse.Schedule(name='2q_schedule')
+        schedule |= const_pulse(system_model.drive(0))  # pi pulse drive
+        schedule += const_pulse(system_model.control(uchannel)) << schedule.duration  # u chan pulse
+
+        acq_cmd = pulse.Acquire(duration=total_samples)
+        schedule |= acq_cmd(system_model.acquires, system_model.memoryslots) << schedule.duration
+
+        return schedule
 
 
 if __name__ == '__main__':
